third_party/blink/web_tests/http/tests/images/feature-policy-unoptimized-lossless-images-expected.txt

CONSOLE ERROR: Feature policy violation: unoptimized-lossless-images is not allowed in this document.
